video_detector/requirements.txt
torch
torchvision
opencv-python
pillow
numpy
pandas
scikit-learn
streamlit
mediapipe
joblib
